Bryter ut overføring til Arena som egen component
