+++ components/overfor-til-arena.js
@@ -1,0 +1,29 @@
+import axios from 'axios'
+import { useState } from 'react'
+import { AlertStripeSuksess, AlertStripeFeil } from 'nav-frontend-alertstriper'
+import { Hovedknapp } from 'nav-frontend-knapper'
+
+const OverforTilArena  = props => {
+  const { id } = props
+  const [status, setStatus] = useState('IKKE_SENDT')
+
+  const handleOverforing = async () => {
+    const { data } = await axios(`${process.env.NEXT_PUBLIC_API_URL}/put-registrering?id=${id}`)
+    const { status } = data
+    setStatus(status)
+  }
+
+  const handleClick = (e) => {
+    e.preventDefault()
+    handleOverforing()
+  }
+  return (
+    <>
+      {status === 'IKKE_SENDT' && <Hovedknapp onClick={handleClick}>Overfør til Arena</Hovedknapp>}
+      {status === 'FEIL' && <AlertStripeFeil>Noe gikk galt under overføringen</AlertStripeFeil>}
+      {status === 'SENDT' && <AlertStripeSuksess>Brukeren er overført</AlertStripeSuksess>}
+    </>
+  )
+}
+
+export default OverforTilArena
